renaming and adding new folders
