+++ digital-intake/frontend/src/main.tsx
@@ -3,7 +3,7 @@ import { createRoot } from "react-dom/client";
 import { Provider } from "react-redux";
 import { RouterProvider } from "react-router-dom";
 import { store } from "./state/store";
-import { router } from "./app/routes";
+import { router } from "./routes/routes";
 import "./index.css";
 
 createRoot(document.getElementById("root")!).render(
